--- src/kernels/variables/IJupyterVariablesProvider.ts
@@ -1,7 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-
-import { NotebookVariableProvider } from 'vscode';
-
-export const IJupyterVariablesProvider = Symbol('IJupyterVariablesProvider');
-export interface IJupyterVariablesProvider extends NotebookVariableProvider {}
